+++ ASimOB/src/org/arosso/sim/TrafficModel.java
@@ -23,25 +23,25 @@ public interface TrafficModel {
 	 * Description of the method getEstimatedPassengerNumber.
 	 * @param time 
 	 */
-	public int getEstimatedPassengerNumber(Long time);
+	public int getEstimatedPassengerNumber(Double time);
 	
 	/**
 	 * Description of the method getEstimatedOriginFloor.
 	 * @param time 
 	 */
-	public int getEstimatedOriginFloor(Long time);
+	public int getEstimatedOriginFloor(Double time);
 	
 	/**
 	 * Description of the method getEstimatedDestinationFloor.
 	 * @param time 
 	 */
-	public int getEstimatedDestinationFloor(Long time);
+	public int getEstimatedDestinationFloor(Double time);
 	
 	/**
 	 * Description of the method getEstimatedArrivalTime.
 	 * @param time 
 	 */
-	public Long getEstimatedArrivalTime(Long time);
+	public Double getEstimatedArrivalTime(Double time);
 	
 	/**
 	 * Description of the method getTrafficMode.
